Tests i TestWinTools virker nu også i x64 Math compilere - _asm blokke

+++ Lib/Include/MFCUtil/AnimatedImage.h
@@ -99,7 +99,7 @@ public:
   void paintAllFrames(const CRect  &r);
   void hide();
   inline int getFrameCount() const {
-    return m_frameTable.size();
+    return (int)m_frameTable.size();
   }
   inline const GifFrame &getFrame(int index) const {
     return m_frameTable[index];
