+++ Lib/Src/GifLib/giflib-5.1.1/lib/gif_font.c
@@ -220,7 +220,7 @@ void GifDrawBoxedText8x8(SavedImage *Image, const int x, const int y, const char
     int leadspace = 0;
 
     if(cp[0] == '\t') {
-      leadspace = (TextWidth - strlen(++cp)) / 2;
+      leadspace = (TextWidth - (int)strlen(++cp)) / 2;
     }
 
     GifDrawText8x8(Image, x + border + (leadspace * GIF_FONT_WIDTH), y + border + (GIF_FONT_HEIGHT * i++), cp, fg);
